Refactor - moved MistmeadowWeitchEffect to shared area
